[#158157958] Move lease to its own service
