+++ app/services/force/application_service.rb
@@ -36,24 +36,8 @@ module Force
       application['proof_files'] = app_proof_files(id) if includes.include?('proof_files')
       application['household_members'] = app_household_members(application) if includes.include?('household_members')
       application['flagged_applications'] = flagged_record_set(id) if includes.include?('flagged_applications')
-      application['lease'] = lease(id) if includes.include?('lease')
+      application['lease'] = lease_service.lease(id) if includes.include?('lease')
       application
-    end
-
-    def lease(application_id)
-      builder.from(:Lease__c)
-             .select(:Id,
-                     :Unit__c,
-                     :Lease_Start_Date__c,
-                     :Lease_Status__c,
-                     :Monthly_Parking_Rent__c,
-                     :Total_Monthly_Rent_without_Parking__c,
-                     :Monthly_Tenant_Contribution__c)
-             .where_eq(:Application__c, application_id, :string)
-             .transform_results { |results| massage(results) }
-             .query
-             .records
-             .first
     end
 
     def listing_applications(listing_id)
@@ -78,42 +62,7 @@ module Force
     end
 
     def submit(data)
-      # If lease information is available, submit it separately.
-      if data.key?(:lease)
-        submit_lease(data[:lease], data[:id], data[:primaryApplicantContact])
-        data = data.except(:lease) # Don't submit lease data to the customAPI.
-      end
       api_post('/LeasingAgentPortal/shortForm', application_defaults.merge(data))
-    end
-
-    def submit_lease(lease, application_id, primary_contact_id)
-      puts 'PRIMARY CONTACT ID', primary_contact_id
-      # Update
-      # FIXME: Fix Tenant__c permissions so we can submit these values.
-      if lease[:id]
-        response = @client.update!('Lease__c',
-                                   Id: lease[:id],
-                                   # Tenant__c: primary_contact_id,
-                                   Unit__c: lease[:unit],
-                                   Lease_Status__c: lease[:leaseStatus],
-                                   Lease_Start_Date__c: lease[:leaseStartDate],
-                                   Monthly_Parking_Rent__c: lease[:monthlyParkingRent],
-                                   Total_Monthly_Rent_without_Parking__c: lease[:totalMonthlyRentWithoutParking],
-                                   Monthly_Tenant_Contribution__c: lease[:monthlyTenantContribution])
-        puts 'Successfully Updated Lease Information: ', response
-      else
-        # Create
-        response = @client.create!('Lease__c',
-                                   Application__c: application_id,
-                                   # Tenant__c: primary_contact_id,
-                                   Unit__c: lease[:unit],
-                                   Lease_Status__c: 'Draft',
-                                   Lease_Start_Date__c: lease[:leaseStartDate],
-                                   Monthly_Parking_Rent__c: lease[:monthlyParkingRent],
-                                   Total_Monthly_Rent_without_Parking__c: lease[:totalMonthlyRentWithoutParking],
-                                   Monthly_Tenant_Contribution__c: lease[:monthlyTenantContribution])
-        puts 'Successfully created new lease: ', response
-      end
     end
 
     private
@@ -181,6 +130,10 @@ module Force
         # for community users, restrict results to their account + draft
         %(Listing__r.Account__c = '#{@user.salesforce_account_id}')
       end
+    end
+
+    def lease_service
+      Force::LeaseService.new(@user)
     end
   end
 end
