+++ app/services/force/lease_service.rb
@@ -1,0 +1,48 @@
+module Force
+  # Encapsulate all Salesforce Lease__c querying and updating functions.
+  class LeaseService < Force::Base
+    def lease(application_id)
+      builder.from(:Lease__c)
+             .select(:Id,
+                     :Unit__c,
+                     :Lease_Start_Date__c,
+                     :Lease_Status__c,
+                     :Monthly_Parking_Rent__c,
+                     :Total_Monthly_Rent_without_Parking__c,
+                     :Monthly_Tenant_Contribution__c)
+             .where_eq(:Application__c, application_id, :string)
+             .transform_results { |results| massage(results) }
+             .query
+             .records
+             .first
+    end
+
+    def submit_lease(lease, application_id, _primary_contact_id)
+      # Update
+      # FIXME: Fix Tenant__c permissions so we can submit these values.
+      if lease[:id]
+        response = @client.update!('Lease__c',
+                                   Id: lease[:id],
+                                   # Tenant__c: primary_contact_id,
+                                   Unit__c: lease[:unit],
+                                   Lease_Status__c: lease[:leaseStatus],
+                                   Lease_Start_Date__c: lease[:leaseStartDate],
+                                   Monthly_Parking_Rent__c: lease[:monthlyParkingRent],
+                                   Total_Monthly_Rent_without_Parking__c: lease[:totalMonthlyRentWithoutParking],
+                                   Monthly_Tenant_Contribution__c: lease[:monthlyTenantContribution])
+      else
+        # Create
+        response = @client.create!('Lease__c',
+                                   Application__c: application_id,
+                                   # Tenant__c: primary_contact_id,
+                                   Unit__c: lease[:unit],
+                                   Lease_Status__c: 'Draft',
+                                   Lease_Start_Date__c: lease[:leaseStartDate],
+                                   Monthly_Parking_Rent__c: lease[:monthlyParkingRent],
+                                   Total_Monthly_Rent_without_Parking__c: lease[:totalMonthlyRentWithoutParking],
+                                   Monthly_Tenant_Contribution__c: lease[:monthlyTenantContribution])
+      end
+      response
+    end
+  end
+end
